[charts] Clean and document Heatmap Tooltip
