+++ packages/x-charts-pro/src/Heatmap/HeatmapTooltip/HeatmapTooltip.classes.ts
@@ -1,0 +1,21 @@
+import composeClasses from '@mui/utils/composeClasses';
+import { getChartsTooltipUtilityClass } from '@mui/x-charts/ChartsTooltip';
+import { HeatmapTooltipProps } from './HeatmapTooltip.types';
+
+export const useUtilityClasses = (props: Pick<HeatmapTooltipProps, 'classes'>) => {
+  const { classes } = props;
+
+  const slots = {
+    root: ['root'],
+    paper: ['paper'],
+    table: ['table'],
+    row: ['row'],
+    cell: ['cell'],
+    mark: ['mark'],
+    markContainer: ['markContainer'],
+    labelCell: ['labelCell'],
+    valueCell: ['valueCell'],
+  };
+
+  return composeClasses(slots, getChartsTooltipUtilityClass, classes);
+};
