+++ packages/x-charts-pro/src/Heatmap/HeatmapTooltip/HeatmapTooltip.tsx
@@ -1,129 +1,18 @@
 'use client';
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import clsx from 'clsx';
 import HTMLElementType from '@mui/utils/HTMLElementType';
-import composeClasses from '@mui/utils/composeClasses';
-import {
-  ChartsTooltipPaper,
-  ChartsTooltipTable,
-  ChartsTooltipRow,
-  ChartsTooltipCell,
-  useItemTooltip,
-  ChartsTooltipContainerProps,
-  getChartsTooltipUtilityClass,
-  ChartsTooltipContainer,
-} from '@mui/x-charts/ChartsTooltip';
-import { useXAxis, useYAxis } from '@mui/x-charts/hooks';
-import { getLabel, ChartsLabelMark } from '@mui/x-charts/internals';
-import { useHeatmapSeriesContext } from '../../hooks/useHeatmapSeries';
-import { HeatmapTooltipAxesValue } from './HeatmapTooltipAxesValue';
-
-export interface HeatmapTooltipProps
-  extends Omit<ChartsTooltipContainerProps, 'trigger' | 'children'> {
-  /**
-   * Select the kind of tooltip to display
-   * - 'item': Shows data about the item below the mouse.
-   * - 'none': Does not display tooltip
-   * @default 'item'
-   */
-  trigger?: 'item' | 'none';
-}
-
-const useUtilityClasses = (props: Pick<HeatmapTooltipProps, 'classes'>) => {
-  const { classes } = props;
-
-  const slots = {
-    root: ['root'],
-    paper: ['paper'],
-    table: ['table'],
-    row: ['row'],
-    cell: ['cell'],
-    mark: ['mark'],
-    markContainer: ['markContainer'],
-    labelCell: ['labelCell'],
-    valueCell: ['valueCell'],
-  };
-
-  return composeClasses(slots, getChartsTooltipUtilityClass, classes);
-};
-
-function DefaultHeatmapTooltipContent(props: Pick<HeatmapTooltipProps, 'classes'>) {
-  const classes = useUtilityClasses(props);
-
-  const xAxis = useXAxis();
-  const yAxis = useYAxis();
-  const heatmapSeries = useHeatmapSeriesContext();
-
-  const tooltipData = useItemTooltip<'heatmap'>();
-
-  if (!tooltipData || !heatmapSeries || heatmapSeries.seriesOrder.length === 0) {
-    return null;
-  }
-
-  const { series, seriesOrder } = heatmapSeries;
-  const seriesId = seriesOrder[0];
-
-  const { color, value, identifier, markType } = tooltipData;
-
-  const [xIndex, yIndex] = value;
-
-  const formattedX =
-    xAxis.valueFormatter?.(xAxis.data![xIndex], {
-      location: 'tooltip',
-      scale: xAxis.scale,
-    }) ?? xAxis.data![xIndex].toLocaleString();
-  const formattedY =
-    yAxis.valueFormatter?.(yAxis.data![yIndex], { location: 'tooltip', scale: yAxis.scale }) ??
-    yAxis.data![yIndex].toLocaleString();
-  const formattedValue = series[seriesId].valueFormatter(value, {
-    dataIndex: identifier.dataIndex,
-  });
-
-  const seriesLabel = getLabel(series[seriesId].label, 'tooltip');
-
-  return (
-    <ChartsTooltipPaper className={classes.paper}>
-      <ChartsTooltipTable className={classes.table}>
-        <HeatmapTooltipAxesValue>
-          <span>{formattedX}</span>
-          <span>{formattedY}</span>
-        </HeatmapTooltipAxesValue>
-        <tbody>
-          <ChartsTooltipRow className={classes.row}>
-            <ChartsTooltipCell className={clsx(classes.labelCell, classes.cell)} component="th">
-              <div className={classes.markContainer}>
-                <ChartsLabelMark type={markType} color={color} className={classes.mark} />
-              </div>
-              {seriesLabel}
-            </ChartsTooltipCell>
-            <ChartsTooltipCell className={clsx(classes.valueCell, classes.cell)} component="td">
-              {formattedValue}
-            </ChartsTooltipCell>
-          </ChartsTooltipRow>
-        </tbody>
-      </ChartsTooltipTable>
-    </ChartsTooltipPaper>
-  );
-}
-
-DefaultHeatmapTooltipContent.propTypes = {
-  // ----------------------------- Warning --------------------------------
-  // | These PropTypes are generated from the TypeScript type definitions |
-  // | To update them edit the TypeScript types and run "pnpm proptypes"  |
-  // ----------------------------------------------------------------------
-  /**
-   * Override or extend the styles applied to the component.
-   */
-  classes: PropTypes.object,
-} as any;
+import { ChartsTooltipContainer } from '@mui/x-charts/ChartsTooltip';
+import { HeatmapTooltipContent } from './HeatmapTooltipContent';
+import { HeatmapTooltipProps } from './HeatmapTooltip.types';
+import { useUtilityClasses } from './HeatmapTooltip.classes';
 
 function HeatmapTooltip(props: HeatmapTooltipProps) {
   const classes = useUtilityClasses({ classes: props.classes });
 
   return (
     <ChartsTooltipContainer trigger="item" {...props} classes={classes}>
-      <DefaultHeatmapTooltipContent classes={classes} />
+      <HeatmapTooltipContent classes={classes} />
     </ChartsTooltipContainer>
   );
 }
