+++ packages/x-charts-pro/src/Heatmap/HeatmapTooltip/HeatmapTooltip.types.ts
@@ -1,0 +1,12 @@
+import { ChartsTooltipContainerProps } from '@mui/x-charts/ChartsTooltip';
+
+export interface HeatmapTooltipProps
+  extends Omit<ChartsTooltipContainerProps, 'trigger' | 'children'> {
+  /**
+   * Select the kind of tooltip to display
+   * - 'item': Shows data about the item below the mouse.
+   * - 'none': Does not display tooltip
+   * @default 'item'
+   */
+  trigger?: 'item' | 'none';
+}
