+++ packages/x-charts-pro/src/Heatmap/HeatmapTooltip/index.ts
@@ -1,1 +1,3 @@
 export * from './HeatmapTooltip';
+export * from './HeatmapTooltipContent';
+export type { HeatmapTooltipProps } from './HeatmapTooltip.types';
