+++ scripts/x-charts-pro.exports.json
@@ -228,6 +228,8 @@
   { "name": "HeatmapClassKey", "kind": "TypeAlias" },
   { "name": "HeatmapPlot", "kind": "Function" },
   { "name": "HeatmapTooltip", "kind": "Function" },
+  { "name": "HeatmapTooltipContent", "kind": "Function" },
+  { "name": "HeatmapTooltipContentProps", "kind": "Interface" },
   { "name": "HeatmapTooltipProps", "kind": "Interface" },
   { "name": "HighlightElementClassKey", "kind": "TypeAlias" },
   { "name": "HighlightItemData", "kind": "TypeAlias" },
